Made `download_scripts` and `utils` - Passed some basic tests - Need a better way to test this
