+++ yt_fts/download_scripts.py
@@ -1,0 +1,200 @@
+import subprocess, re, os, sqlite3, json
+
+from progress.bar import Bar
+from concurrent.futures import ThreadPoolExecutor
+from bs4 import BeautifulSoup
+
+from yt_fts.db_scripts import add_video
+
+
+def get_vid_title(vid_url, s):
+    res = s.get(vid_url)
+    if res.status_code == 200:
+        html = res.text
+        soup = BeautifulSoup(html, 'html.parser')
+        title = soup.title.string
+        return title 
+    else:
+        return None
+
+
+def get_videos_list(channel_url):
+    cmd = [
+        "yt-dlp",
+        "--flat-playlist",
+        "--print",
+        "id",
+        f"{channel_url}"
+    ]
+    res = subprocess.run(cmd, capture_output=True, check=True)
+    list_of_videos_urls = res.stdout.decode().splitlines()
+    return list_of_videos_urls 
+
+
+def get_channel_id(url, s):
+    res = s.get(url)
+    if res.status_code == 200:
+        html = res.text
+        channel_id = re.search('channelId":"(.{24})"', html).group(1)
+        print(channel_id)
+        return channel_id
+    else:
+        return None
+
+
+def get_channel_name(channel_id, s):
+
+    res = s.get(f"https://www.youtube.com/channel/{channel_id}/videos")
+
+    if res.status_code == 200:
+
+        html = res.text
+        soup = BeautifulSoup(html, 'html.parser')
+        script = soup.find('script', type='application/ld+json')
+
+        # Hot fix for channels with special characters in the name
+        try:
+            print("Trying to parse json normally")
+            data = json.loads(script.string)
+        except:
+            print("json parse failed retrying with escaped backslashes")
+            script = script.string.replace('\\', '\\\\')
+            data = json.loads(script)
+
+        channel_name = data['itemListElement'][0]['item']['name']
+        print(channel_name)
+        return channel_name 
+    else:
+        return None
+
+
+def handle_reject_consent_cookie(channel_url, s):
+    r = s.get(channel_url)
+    if "https://consent.youtube.com" in r.url:
+        m = re.search(r"<input type=\"hidden\" name=\"bl\" value=\"([^\"]*)\"", r.text)
+        if m:
+            data = {
+                "gl":"DE",
+                "pc":"yt",
+                "continue":channel_url,
+                "x":"6",
+                "bl":m.group(1),
+                "hl":"de",
+                "set_eom":"true"
+            }
+            s.post("https://consent.youtube.com/save", data=data)
+
+
+
+def download_vtts(number_of_jobs, list_of_videos_urls, language ,tmp_dir):
+    executor = ThreadPoolExecutor(number_of_jobs)
+    futures = []
+    for video_id in list_of_videos_urls:
+        video_url = f'https://www.youtube.com/watch?v={video_id}'
+        future = executor.submit(get_vtt, tmp_dir, video_url, language)
+        futures.append(future)
+    
+    for i in range(len(list_of_videos_urls)):
+        futures[i].result()
+
+
+def get_vtt(tmp_dir, video_url, language):
+    subprocess.run([
+        "yt-dlp",
+        "-o", f"{tmp_dir}/%(id)s.%(ext)s",
+        "--write-auto-sub",
+        "--convert-subs", "vtt",
+        "--skip-download",
+        "--sub-langs", f"{language},-live_chat",
+        video_url
+    ])
+
+
+def download_vtts(number_of_jobs, list_of_videos_urls, language ,tmp_dir):
+    executor = ThreadPoolExecutor(number_of_jobs)
+    futures = []
+    for video_id in list_of_videos_urls:
+        video_url = f'https://www.youtube.com/watch?v={video_id}'
+        future = executor.submit(get_vtt, tmp_dir, video_url, language)
+        futures.append(future)
+    
+    for i in range(len(list_of_videos_urls)):
+        futures[i].result()
+
+
+def get_vtt(tmp_dir, video_url, language):
+    subprocess.run([
+        "yt-dlp",
+        "-o", f"{tmp_dir}/%(id)s.%(ext)s",
+        "--write-auto-sub",
+        "--convert-subs", "vtt",
+        "--skip-download",
+        "--sub-langs", f"{language},-live_chat",
+        video_url
+    ])
+
+
+
+def vtt_to_db(channel_id, dir_path, s):
+    items = os.listdir(dir_path)
+
+    file_paths = []
+
+    for item in items:
+        item_path = os.path.join(dir_path, item)
+        if os.path.isfile(item_path):
+            file_paths.append(item_path)    
+
+    con = sqlite3.connect('subtitles.db')  
+    cur = con.cursor()
+
+    bar = Bar('Processing', max=len(file_paths))
+
+    for vtt in file_paths:
+        base_name = os.path.basename(vtt)
+        vid_id = re.match(r'^([^.]*)', base_name).group(1)
+        vid_url = f"https://youtu.be/{vid_id}"
+        vid_title = get_vid_title(vid_url, s)
+        add_video(channel_id, vid_id, vid_title, vid_url)
+
+        vtt_json = parse_vtt(vtt)
+
+        for sub in vtt_json:
+            start_time = sub['start_time']
+            text = sub['text']
+            cur.execute(f"INSERT INTO Subtitles (video_id, timestamp, text) VALUES (?, ?, ?)", (vid_id, start_time, text))
+
+        con.commit()
+        bar.next()
+
+    bar.finish() 
+    con.close()
+
+
+def parse_vtt(file_path):
+
+    result = []
+
+    time_pattern = "^(.*) align:start position:0%"
+
+    with open(file_path, "r") as f:
+        lines = f.readlines()
+
+    for count, line in enumerate(lines):
+        time_match = re.match(time_pattern, line)
+
+        if time_match:
+            start = re.search("^(.*) -->",time_match.group(1))
+            start_time = start.group(1)
+            sub_titles = lines[count + 1]
+
+            # prevent duplicate entries
+            if result and result[-1]['text'] == sub_titles.strip('\n'):
+                continue
+            else:   
+                result.append({
+                    'start_time': start_time,
+                    'text': sub_titles.strip('\n'),
+                })
+
+    return result
